backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiosmtplib==3.0.1
email-validator==2.2.0
psycopg[binary]==3.2.1
alembic==1.13.2
redis==5.0.8
rq==1.16.2
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
requests==2.32.3
